Sort actor nodes list in axon-tui Add actor details view to axon-tui (YAML format)

+++ axon-tui/ui/mainWidget.go
@@ -9,18 +9,23 @@ import (
 // MainWidget represents the main widget of the application
 type MainWidget struct {
 	*ui.Block
+	Control
 	header    *HeaderWidget
 	nodesGrid *NodesGridWidget
 	eventsCh  chan ui.Event
-	visible   bool
-	inFocus   bool
 }
 
 // Draw draws the content of the widget into the `Buffer` object provided as an argument
 func (m *MainWidget) Draw(buf *ui.Buffer) {
-	ui.Clear()
-	m.Block.Draw(buf)
-	ui.Render(m.header, m.nodesGrid)
+	if m.Visible {
+		ui.Clear()
+		m.header.Visible = true
+		m.nodesGrid.Visible = true
+		ui.Render(m.header, m.nodesGrid)
+	} else {
+		m.header.Visible = false
+		m.nodesGrid.Visible = false
+	}
 }
 
 // NewMainWidget create a new instance of the main widget of the application
@@ -34,8 +39,6 @@ func NewMainWidget(epnStatus *epn.Status, eventsHub *EventsHub) *MainWidget {
 		header:    NewHeaderWidget(termWidth, headerHeight),
 		nodesGrid: NewNodesGridWidget(termWidth, termHeight, headerHeight, epnStatus, eventsHub),
 		eventsCh:  eventsHub.Subscribe(),
-		visible:   true,
-		inFocus:   false,
 	}
 
 	main.Border = false
@@ -48,15 +51,17 @@ func NewMainWidget(epnStatus *epn.Status, eventsHub *EventsHub) *MainWidget {
 // controller is the event handler and controller of the widget
 func (m *MainWidget) controller() {
 	for e := range m.eventsCh {
-		switch e.ID {
-		case "q", "<C-c>":
-			if err := syscall.Kill(syscall.Getpid(), syscall.SIGTERM); err != nil {
-				panic(err)
-			}
-			return
+		if m.UseEvents {
+			switch e.ID {
+			case "q", "<Escape>", "<C-c>":
+				if err := syscall.Kill(syscall.Getpid(), syscall.SIGTERM); err != nil {
+					panic(err)
+				}
+				return
 
-		case "<Resize>":
-			m.resize(e.Payload.(ui.Resize))
+			case "<Resize>":
+				m.resize(e.Payload.(ui.Resize))
+			}
 		}
 	}
 }
@@ -66,6 +71,5 @@ func (m *MainWidget) resize(size ui.Resize) {
 	m.header.SetRect(0, 0, size.Width, m.header.Height)
 	m.nodesGrid.SetRect(0, m.header.Height, size.Width, size.Height)
 
-	ui.Clear()
-	ui.Render(m.header, m.nodesGrid)
+	ui.Render(m)
 }
